fastapi_processor/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.7
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.25
python-jose[cryptography]==3.3.0
openai==1.10.0
httpx==0.26.0
python-decouple==3.8
websockets==12.0
aiofiles==23.2.1
Pillow==10.2.0
PyPDF2==3.0.1
python-docx==1.1.0
pydantic==2.5.3
pydantic-settings==2.1.0
